refactor old code to make it more reusable and responsive - add wishlist container
