+++ src/components/shared/Counter/Counter.jsx
@@ -33,7 +33,6 @@ function Counter() {
         minWidth:'20px',
         alignItems: "center",
         padding: "0 5px",
-        margin: "0 10px",
         border: `1px solid ${theme.palette.primary.main}`,
         borderRadius: "4px",
         color: `${theme.palette.primary.main}`,
